requirements.txt
google-cloud-vision
google-cloud-storage
tqdm
